Fase 2 con word en marcha

+++ storage/framework/textos/c0c9f4b2edd70f4b49093c0a9096900f92cc0c35.php.php
@@ -5,6 +5,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Controllers\ImageController;
 use App\Http\Controllers\StorageController;
 use App\User;
+use App\Autor;
 use App\Role;
 use App\Imagenes;
 use App\ImagenesRelacion;
@@ -15,34 +16,47 @@ use Yajra\Datatables\Datatables;
 class UserController extends Controller
 {
     public  $user;
-     public function index(Request $request){
-         $data = User::orderBy('lname','ASC')->paginate(25);
+    public function index(Request $request){
+        $data = User::orderBy('lname','ASC')->paginate(25);
         return view('users.index',compact('data'))
             ->with('i', ($request->input('page', 1) - 1) * 5);
-     }
+    }
 
     public function create()
     {
+
         $roles = Role::pluck('display_name','id');
+
         return view('users.create',compact('roles'));
     }
 
     public function store(Request $request)
     {
         $this->validate($request, [
-            'name' => 'required',
-            'email' => 'required|email|unique:users,email',
+            'name'     => 'required',
+            'email'    => 'required|email|unique:users,email',
             'password' => 'required|same:confirm-password',
-            'roles' => 'required'
-        ]);
+            'roles'    => 'required'
+        ]);//'ldap'     => 'required|unique:users,ldap',
         $input = $request->all();
         $input['password'] = Hash::make($input['password']);
+        $input['name']  = ucfirst(strtolower(trim($input['name'])));
+        $input['lname'] = ucfirst(strtolower(trim($input['lname'])));
+
         $user = User::create($input);
         foreach ($request->input('roles') as $key => $value) {
             $user->attachRole($value);
         }
+        $valores = [
+           'user_id'  => $user->id,
+           'nombre'   => $user->name,
+           'apellido' => $user->lname
+        ];
+        $autor = Autor::create($valores);
+        $valorUpdate['id_autor'] = $autor->id;
+        $user->update($valorUpdate);
         return redirect()->route('users.index')
-                        ->with('success','User created successfully');
+            ->with('success', __('Erabiltzailea zuzen sortu da'));
     }
 
     public function show($id)
@@ -53,13 +67,8 @@ class UserController extends Controller
     public function edit($id)
     {
         $user = User::find($id);
-
         $roles = Role::pluck('display_name','id');
         $userRole = $user->roles->pluck('id','id')->toArray();
-        $imagenes = new ImageController();
-        $imagenes = $imagenes->dameImagenes( 'user_', $id, '1');
-        $adjuntos = new StorageController();
-        $adjuntos = $adjuntos->dameAdjuntos( 'user_', $id);
 
         return view('users.edit',compact('user','roles','userRole', 'imagenes', 'adjuntos'));
     }
@@ -67,12 +76,16 @@ class UserController extends Controller
     public function update(Request $request, $id)
     {
         $this->validate($request, [
-            'name' => 'required',
-            'email' => 'required|email|unique:users,email,'.$id,
-            'password' => 'same:confirm-password',
-            'roles' => 'required'
-        ]);
-        $input = $request->all();
+            'name'     => 'required',
+            'lname'    => 'required',
+            'email'    => 'required|email|unique:users,email,'.$id,
+            'password' => 'same:confirm-password'
+        ]);// 'ldap'     => 'required',
+
+        $input          = $request->all();
+        $input['name']  = ucfirst(strtolower(trim($input['name'])));
+        $input['lname'] = ucfirst(strtolower(trim($input['lname'])));
+        $input['estado'] = '1';
         if(!empty($input['password'])){
             $input['password'] = Hash::make($input['password']);
         }else{
@@ -80,22 +93,41 @@ class UserController extends Controller
         }
         $user = User::find($id);
         $user->update($input);
-
-
-
-        DB::table('role_user')->where('user_id',$id)->delete();
-
-        foreach ($request->input('roles') as $key => $value) {
-            $user->attachRole($value);
+        if(!empty($request->input('roles'))){
+            DB::table('role_user')->where('user_id',$id)->delete();
+            foreach ($request->input('roles') as $key => $value) {
+                $user->attachRole($value);
+            }
         }
-        return redirect()->route('users.index')->with('success','User updated successfully');
+        if ( empty($user->autor) OR ($user->autor == '0') ){
+           $valores = [
+                'user_id'   => $id,
+                'nombre'    => ucfirst(strtolower(trim($user->name))),
+                'apellido'  => ucfirst(strtolower(trim($user->lname))),
+                'tipo'      => 'EHU'
+            ];
+            $autor = Autor::create($valores);
+            $input['id_autor'] = $autor->id;
+            $user->update($input);
+        }else{
+            $autor = Autor::find($user->autor->id);
+            $valores = [
+               'nombre'  => ucfirst(strtolower(trim($user->name))),
+               'apellido'  => ucfirst(strtolower(trim($user->lname))),
+
+            ];
+            $autor->update($valores);
+            $input['id_autor'] = $autor->id;
+            $user->update($input);
+        }
+        return redirect()->route('users.index')->with('success', __('Erabiltzailea zuzen aldatatu da'));
     }
 
     public function destroy($id)
     {
         User::find($id)->delete();
         return redirect()->route('users.index')
-                        ->with('success','User deleted successfully');
+                        ->with('success', __('Erabiltzailea zuzen ezabatu da'));
     }
     public static function getUserName($user_id)
     {
